fix: 検索処理をクラス分割
